+++ src/frontend-cdoc-commons/resolver/cdoc-searchform.resolver.ts
@@ -2,7 +2,9 @@ import {ActivatedRouteSnapshot, Resolve, RouterStateSnapshot} from '@angular/rou
 import {AppState, GenericAppService} from '@dps/mycms-commons/dist/commons/services/generic-app.service';
 import {ResolvedData, ResolverError} from '../../angular-commons/resolver/resolver.utils';
 import {CommonDocSearchForm} from '@dps/mycms-commons/dist/search-commons/model/forms/cdoc-searchform';
-import {GenericSearchFormSearchFormConverter} from '@dps/mycms-commons/dist/search-commons/services/generic-searchform.converter';
+import {
+    GenericSearchFormSearchFormConverter
+} from '@dps/mycms-commons/dist/search-commons/services/generic-searchform.converter';
 
 export abstract class CommonDocSearchFormResolver<F extends CommonDocSearchForm> implements Resolve<ResolvedData<F>> {
     static ERROR_INVALID_SEARCHFORM = 'ERROR_INVALID_SEARCHFORM';