+++ src/frontend-pdoc-module/pdoc-admin/components/pdoc-editpage/pdoc-editpage.component.spec.ts
@@ -12,7 +12,6 @@ import {GenericAppService} from '@dps/mycms-commons/dist/commons/services/generi
 import {ErrorResolver} from '../../../../frontend-cdoc-commons/resolver/error.resolver';
 import {PageUtils} from '../../../../angular-commons/services/page.utils';
 import {AngularMarkdownService} from '../../../../angular-commons/services/angular-markdown.service';
-import {NgxMdModule, NgxMdService} from 'ngx-md';
 import {AngularHtmlService} from '../../../../angular-commons/services/angular-html.service';
 import {CommonRoutingService} from '../../../../angular-commons/services/common-routing.service';
 import {RouterStub} from '../../../../angular-commons/testing/router-stubs';
@@ -41,7 +40,6 @@ describe('PDocEditpageComponent', () => {
             declarations: [PDocEditpageComponent],
             imports: [
                 TranslateModule.forRoot(),
-                NgxMdModule.forRoot(),
                 NoopAnimationsModule
             ],
             providers: [
@@ -56,7 +54,6 @@ describe('PDocEditpageComponent', () => {
                 PDocRoutingService,
                 { provide: ToastrService, useValue: new ToastrServiceStub() },
                 TranslateService,
-                NgxMdService,
                 AngularMarkdownService,
                 AngularHtmlService,
                 ErrorResolver,