+++ src/markdown-commons/extensions/markdown.extensions.ts
@@ -1,0 +1,10 @@
+import {MarkdownExtension} from './markdown.extension';
+import {MarkdownDescriptionExtension, MarkdownDescriptionListExtension} from './markdown.dl.extension';
+import {MarkdownRuleBoxEndExtension, MarkdownRuleBoxStartExtension} from './markdown.rulebox.extension';
+
+export const MarkdownDefaultExtensions: MarkdownExtension[] = [
+    MarkdownDescriptionListExtension,
+    MarkdownDescriptionExtension,
+    MarkdownRuleBoxStartExtension,
+    MarkdownRuleBoxEndExtension
+];